Use Reply-To instead of spoofing email addresses
At some point recently, the mailservers responsible for getting
mail from our webserver to our ticketing system started being
particular about checking for spoofed From addresses. So two
changes are made to how the emails are now constructed:

1. They are "From" whatever address mailx is set to use by default
(www-data@local-hostname.whatever), which the mailservers are
happy with since it's not a spoofed address
2. They have a "Reply-To" address matching the preferred email
given in the form, which the ticketing system is happy to use
as the Requestor field.