Adjusting tabs for file structure example
It's not clear to me what exactly the file structure is supposed to be, but the inconsistency is apparent.  I believe that `job_input` is supposed to be within the `input/` directory, and that in turn is within the `job#/` directory.  That way the `job_input` is transferred for each job submitted per `job#/` directory, per the corresponding submit file.